+++ src/nacl.rs
@@ -2,6 +2,7 @@
 
 /// Extension ID for Nested Acceleration Extension.
 pub const EID_NACL: usize = crate::eid_from_str("NACL") as _;
+
 pub use fid::*;
 
 /// Declared in § 15.15.
@@ -48,4 +49,27 @@ pub mod feature_id {
     ///
     /// Declared in §15.4.
     pub const AUTOSWAP_CSR: usize = 3;
+}
+
+/// Size of shared memory set by supervisor software for current hart.
+///
+/// NACL shared memory includes scratch space and CSR space. Due to the difference
+/// of CSR width, this size varies between different `XLEN` values. `NATIVE`
+/// constant here only matches the integer width for the target this crate is compiled.
+/// If you are writing an SEE with different `XLEN` from host platform, you should
+/// choose other correct constant value from `RV32`, `RV64` or `RV128` in module `shmem_size`
+/// instead.
+pub mod shmem_size {
+    use core::mem::size_of;
+    /// Size of NACL shared memory on platforms with `XLEN` of same width as the current platform.
+    pub const NATIVE: usize = 4096 + 1024 * size_of::<usize>();
+
+    /// Size of NACL shared memory on RV32 platforms.
+    pub const RV32: usize = 4096 + 1024 * size_of::<u32>();
+
+    /// Size of NACL shared memory on RV64 platforms.
+    pub const RV64: usize = 4096 + 1024 * size_of::<u64>();
+
+    /// Size of NACL shared memory on RV128 platforms.
+    pub const RV128: usize = 4096 + 1024 * size_of::<u128>();
 }